chore: moves condition-related to logic to its own file
